Интеграционное тестирование

+++ app/controllers/questions_controller.rb
@@ -1,19 +1,23 @@
 class QuestionsController < ApplicationController
+  before_action :authenticate_user!, except: %i[index show]
+
   def index
     @questions = Question.all
   end
 
-  def show; end
+  def show
+    @answer ||= question.answers.new
+  end
 
   def new; end
 
   def edit; end
 
   def create
-    @question = Question.new(question_params)
+    @question = current_user.questions.new(question_params)
 
     if @question.save
-      redirect_to @question
+      redirect_to @question, notice: 'Your question successfully created.'
     else
       render :new
     end
@@ -28,8 +32,8 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    question.destroy if current_user.author?(question)
+    redirect_to questions_path, notice: 'Your question was successfully deleted!'
   end
 
   private
